New things and squahed bugs
Added the grey inactive panel over pages. Also, set a original size for
the server. The Jtree might be smaller than hoped for, but I'm trying
to think of a way to fix it without hardcoding the size.
Changed the PagingGUI to a JRootPane for the inactive pane.
That's about it :)